Fix some YouTube videos and delete dead links
For some reason, some of the YouTube videos were marked as private and as such
didn't load, therefore I just deleted them, but later found the originals on a
different Google account and changed them to unlisted and public and now they
work, so I've brought back the embedded videos to the posts.

All except one, which just says it's not available and I'm not sure why, so that
was just deleted.

Also:

Day 4 of Japan 2010 must have had the original images posted with localhost URLs
and therefore do not work. This isn't from the migration process as the original
is the same:
https://japan2010.stevenocchipinti.com/2010/01/day-4-akihabara.html

So I just deleted them.