Improve ordering of trips based on startDate
This gives me a better field to sort the trips by instead of just a year so when
multiple trips within one year occur, they still appear in order.

I also changed the names of the files to be based on the starting year rather
than just where the majority of the days were to be more consistent.

Closes #7