Support browsers that don't support animation-timeline
Without the @supports, browsers like Safari and Firefox only render the final state of the animation, which is a black box instead of the image.

I also used the svh unit on the cover image to prevent it from changing value on mobile browsers when the browser chrome is visible.