Upgrade Astro to use inferSize for images
I thought this would cause a CLS issue, but according to lighthouse it doesn't
make much difference, maybe due to the blur-up I'm using as that small image
comes in pretty quick.

When I slow down the network or block cloudinary requests, the non-grid images
do start completely collapsed before this change, and have the correct height
with this change, so I may as well keep it in, unless I find any real downside
later.