Un-nest the prose related CSS
This wasn't working in Samsung Internet. I assumed Astro was transpiling this,
but it mustn't have and while it worked in Chrome (which Samsung Internet is
based on), it didn't work in Samsung Internet but this change made it work.

The other option was to include SASS, if this file gets any more complex,
probably worth bringing it in, even just for the nesting.